pyproject.toml
[project]
name = "veterinary-clinic"
version = "0.1.0"
description = "Managing appointments in a veterinary clinic project"
authors = [
    {name = "Mikita Tsybulka",email = "nikita.tsybulka@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "django-jet-reboot (>=1.3.10,<2.0.0)",
    "django-simple-captcha (>=0.6.2,<0.7.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
